refactor: extract finetune run helpers and benchmark rule into focused modules
