+++ data/feature_index.json
@@ -1179,6 +1179,6 @@
       ]
     }
   ],
-  "generated_at": "2026-05-31T15:42:34.681028Z",
+  "generated_at": "2026-05-31T17:08:28.649411Z",
   "schema_version": 1
 }
