+++ scripts/_ghost_wiring_manifest.txt
@@ -125,4 +125,4 @@ ENFORCED CockpitService #1981 -- constructed in api/app.py::_wire_cockpit_servic
 ENFORCED FlightRecorderService #1981 -- constructed in api/app.py::_wire_cockpit_services; frame-authoritative get/seek behind the flight-recorder cockpit endpoints
 ENFORCED build_steering_directive #1981 -- called in api/app.py::_wire_cockpit_services; builds the SafeDefaultSteeringDirective that delivers cockpit hint/redirect as INFO_REQUEST interrupts
 ENFORCED build_flight_recorder_sink #1981 -- called in workers/runtime_builder._build_flight_recorder_sink; selects the recorder sink the boot AgentEngine records per-turn frames through
-ENFORCED FineTuneOrchestrator #1990 -- constructed in api/lifecycle_helpers/feature_wiring.py::_wire_fine_tune_orchestrator behind a connected persistence backend; owns the embedding finetune pipeline and (with a memory backend) a TrajectoryTrainingDataSource so a trajectory-mode run sources real org history
+ENFORCED FineTuneOrchestrator #1990 -- constructed in api/lifecycle_helpers/finetune_wiring.py::_wire_fine_tune_orchestrator behind a connected persistence backend; owns the embedding finetune pipeline and (with a memory backend) a TrajectoryTrainingDataSource so a trajectory-mode run sources real org history
