+++ src/synthorg/meta/models.py
@@ -807,7 +807,6 @@ from synthorg.meta.signal_models import (  # noqa: E402
     ErrorCategorySummary,
     EvolutionOutcomeSummary,
     MetricSummary,
-    OrgBenchmarkSummary,
     OrgBudgetSummary,
     OrgCoordinationSummary,
     OrgErrorSummary,
@@ -838,7 +837,6 @@ __all__ = [
     "ImprovementCycleResult",
     "ImprovementProposal",
     "MetricSummary",
-    "OrgBenchmarkSummary",
     "OrgBudgetSummary",
     "OrgCoordinationSummary",
     "OrgErrorSummary",
