+++ src/synthorg/meta/rules/builtin.py
@@ -17,6 +17,7 @@ from synthorg.meta.models import (
     RuleMatch,
     RuleSeverity,
 )
+from synthorg.meta.rules.benchmark_rule import BenchmarkRegressionRule
 from synthorg.observability import get_logger
 
 if TYPE_CHECKING:
@@ -34,9 +35,6 @@ _DEFAULT_REDUNDANCY_THRESHOLD: Final[float] = 0.3
 _DEFAULT_SCALING_FAILURE_THRESHOLD: Final[float] = 0.5
 _DEFAULT_SCALING_MIN_DECISIONS: Final[int] = 3
 _DEFAULT_ERROR_SPIKE_THRESHOLD: Final[int] = 10
-# A regression needs a predecessor run to compare against, so the rule
-# only fires once at least this many runs are on the curve.
-_BENCHMARK_REGRESSION_MIN_RUNS: Final[int] = 2
 
 # ── Performance rules ──────────────────────────────────────────────
 
@@ -587,69 +585,6 @@ class ErrorSpikeRule:
                 suggested_altitudes=self.target_altitudes,
             )
         return None
-
-
-# ── Benchmark rules ────────────────────────────────────────────────
-
-
-class BenchmarkRegressionRule:
-    """Fires CRITICAL when the latest golden-benchmark run regressed.
-
-    The golden-company benchmark is the org's ground-truth quality
-    signal, so a regression -- the latest scored run dropping materially
-    below its predecessor -- is the strongest "something got worse"
-    signal available and warrants the highest severity. It suggests
-    prompt-tuning and code-modification remediations, the altitudes that
-    can move a benchmark score back up.
-    """
-
-    @property
-    def name(self) -> NotBlankStr:
-        """Rule name.
-
-        Returns:
-            ``NotBlankStr`` instance.
-        """
-        return NotBlankStr("benchmark_regression")
-
-    @property
-    def target_altitudes(self) -> tuple[ProposalAltitude, ...]:
-        """Suggests prompt tuning and code modification.
-
-        Returns:
-            Tuple of the declared element types.
-        """
-        return (
-            ProposalAltitude.PROMPT_TUNING,
-            ProposalAltitude.CODE_MODIFICATION,
-        )
-
-    def evaluate(self, snapshot: OrgSignalSnapshot) -> RuleMatch | None:
-        """Check whether the latest benchmark run regressed.
-
-        Returns:
-            The ``RuleMatch`` value when present, ``None`` otherwise.
-        """
-        bench = snapshot.benchmark
-        if bench.run_count < _BENCHMARK_REGRESSION_MIN_RUNS:
-            return None
-        if not bench.is_regression:
-            return None
-        return RuleMatch(
-            rule_name=self.name,
-            severity=RuleSeverity.CRITICAL,
-            description=(
-                f"Benchmark score dropped {abs(bench.delta)} points "
-                f"(latest {bench.latest_total}/{bench.max_total})"
-            ),
-            signal_context={
-                "latest_total": bench.latest_total,
-                "max_total": bench.max_total,
-                "delta": bench.delta,
-                "run_count": bench.run_count,
-            },
-            suggested_altitudes=self.target_altitudes,
-        )
 
 
 # ── Default rule set ───────────────────────────────────────────────
